fix(infinite-scroll): pr comments
