+++ packages/x-components/src/utils/html.ts
@@ -12,10 +12,10 @@ export function isElementEqualOrContained(a: Element, b: Element): boolean {
 }
 
 /**
- * Returns the target element for a given event. The target element is got form `composedPath` Event
- * method because if the event is triggered inside a Shadow DOM context, `event.target` points to
- * Shadow DOM element, not the element that triggered the event. `composedPath` method also is
- * available in a non-shadow DOM context. `event.target` is added as fallback.
+ * Returns the target element for a given event. The target element is obtained from `composedPath`
+ * Event method because if the event is triggered inside a Shadow DOM context, `event.target` points
+ * to Shadow DOM element, not the element that triggered the event. `composedPath` method also is
+ * available in a non-shadow DOM context.
  *
  * @remarks In a shadow DOM context, this function only works if the Shadow DOM uses `open`
  * encapsulation mode.
@@ -26,5 +26,5 @@ export function isElementEqualOrContained(a: Element, b: Element): boolean {
  * @public
  */
 export function getTargetElement(event: Event): Element {
-  return (event.composedPath()[0] ?? event.target) as Element;
+  return event.composedPath()[0] as Element;
 }
